+++ esibd/devices/__init__.py
@@ -9,5 +9,6 @@ from .omnicontrol.omnicontrol import OMNICONTROL  # noqa: F401
 from .pico.pico import PICO  # noqa: F401
 from .rbd.rbd import RBD  # noqa: F401
 from .rspd3303c.rspd3303c import RSPD3303C  # noqa: F401
+from .spa.spa import SPA1x0  # noqa: F401
 from .temperature.temperature import Temperature  # noqa: F401
 from .tic.tic import TIC  # noqa: F401